+++ bootstrap/providers.php
@@ -7,6 +7,7 @@ return [
     App\Core\Roles\Providers\RoleServiceProvider::class,
     App\Core\Settings\Providers\SettingsServiceProvider::class,
     App\Core\Tooling\Providers\ToolingServiceProvider::class,
+    App\Core\Translations\Providers\CoreTranslationsServiceProvider::class,
     App\Core\Users\Providers\UserServiceProvider::class,
     App\Providers\AppServiceProvider::class,
     App\Providers\ExtensionServiceProvider::class,